Re-add the menpofit logic
When menpofit is missing, menpowidgets should not try
and import it's methods. This was previously in the logic,
and now it is there again. It looks like a lot of changes but
it is mostly due to moving functions around. I have left
appearance model and shape model viewing public.