Fix new observe bugs from IPywidgets 4.1
It seems that on_trait_change has been deprecated and the
implementation of observing and unobserving has become
much more strict. This makes sure that we only observe and
unobserve when neccessary (for the Exceptions thrown by the
visualize_images widget). Other widgets may also be broken.
We also need to stop using on_trait_change and possibly
investigate if the widgets are much slower now.